requirements.txt
anyio==4.2.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.11.17
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
python-dotenv==1.0.0
python-telegram-bot==20.7
pytz==2023.3.post1
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
tornado==6.4
tzlocal==5.2
urllib3==1.26.15
wheel==0.42.0
